Create block device mapping for utility instances
Add a block_device_mapping for the utility instances to provide a disk with 32GB of space. The default instance disk space is too small to hold a Stage database dump with migrated data.

Once this is merged and the build finishes, an SRE will need to terminate the existing utility instance so AWS will deploy a fresh one with the new block storage.